refactor(encoder.c): move encoder code to second file
